Add fault tolerance configuration for pipelines.
This allows a pipeline to be configured as either fault tolerant or not,
and then controller startup enforces that.  This also adapts the
pipeline manager to keep the pipeline directory around if fault
tolerance is configured (but delete the file that indicates the port
number before startup, so that we don't think a later process is using
an earlier one's HTTP listening port).

+++ crates/pipeline-manager/src/runner/local_runner.rs
@@ -9,20 +9,17 @@ use crate::runner::error::RunnerError;
 use crate::runner::logs_buffer::LogsBuffer;
 use crate::runner::pipeline_executor::{LogMessage, PipelineExecutor};
 use async_trait::async_trait;
-use feldera_types::config::{PipelineConfig, StorageCacheConfig, StorageConfig};
+use feldera_types::config::{FtConfig, PipelineConfig, StorageCacheConfig, StorageConfig};
 use log::{debug, error, info};
 use std::path::Path;
 use std::process::Stdio;
 use std::time::Duration;
+use tokio::fs::{remove_dir_all, remove_file};
 use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
 use tokio::process::{Child, ChildStderr, ChildStdout, Command};
 use tokio::sync::{mpsc, oneshot};
 use tokio::task::JoinHandle;
-use tokio::{
-    fs,
-    fs::{create_dir_all, remove_dir_all},
-    select, spawn,
-};
+use tokio::{fs, fs::create_dir_all, select, spawn};
 
 /// Retrieve the binary executable using its URL.
 pub async fn fetch_binary_ref(
@@ -130,6 +127,7 @@ pub struct LocalRunner {
         oneshot::Sender<()>,
         JoinHandle<mpsc::Receiver<mpsc::Sender<LogMessage>>>,
     )>,
+    delete_pipeline_dir_on_shutdown: bool,
 }
 
 impl Drop for LocalRunner {
@@ -305,6 +303,7 @@ impl PipelineExecutor for LocalRunner {
                 log_reject_terminate_sender,
                 log_reject_join_handle,
             )),
+            delete_pipeline_dir_on_shutdown: false,
         }
     }
 
@@ -339,9 +338,22 @@ impl PipelineExecutor for LocalRunner {
         program_binary_url: &str,
         program_version: Version,
     ) -> Result<(), ManagerError> {
-        // (Re-)create pipeline working directory (which will contain storage directory)
+        // Create pipeline directory and write metadata and config files to it.
         let pipeline_dir = self.config.pipeline_dir(self.pipeline_id);
-        let _ = remove_dir_all(&pipeline_dir).await;
+        let ft = deployment_config.global.fault_tolerance;
+        match ft {
+            None | Some(FtConfig::InitialState) => {
+                let _ = remove_dir_all(&pipeline_dir).await;
+            }
+            Some(FtConfig::LatestCheckpoint) => {
+                // Delete the port file because otherwise an old one can make us
+                // think the new pipeline program is listening on an old port.
+                let port_file_path = self.config.port_file_path(self.pipeline_id);
+                let _ = remove_file(port_file_path).await;
+            }
+        };
+        self.delete_pipeline_dir_on_shutdown = ft.is_none();
+        // (Re-)create pipeline working directory (which will contain storage directory)
         create_dir_all(&pipeline_dir).await.map_err(|e| {
             ManagerError::io_error(
                 format!(
@@ -515,14 +527,16 @@ impl PipelineExecutor for LocalRunner {
         }
 
         // Remove the pipeline working directory
-        match remove_dir_all(self.config.pipeline_dir(self.pipeline_id)).await {
-            Ok(_) => (),
-            Err(e) => {
-                log::warn!(
-                    "Failed to delete pipeline working directory for pipeline {}: {}",
-                    self.pipeline_id,
-                    e
-                );
+        if self.delete_pipeline_dir_on_shutdown {
+            match remove_dir_all(self.config.pipeline_dir(self.pipeline_id)).await {
+                Ok(_) => (),
+                Err(e) => {
+                    log::warn!(
+                        "Failed to delete pipeline working directory for pipeline {}: {}",
+                        self.pipeline_id,
+                        e
+                    );
+                }
             }
         }
